cleanup: split catalogue snapshot activation
